ika_web/requirements.txt
aniso8601==8.0.0
apispec==3.3.2
apispec-webframeworks==0.5.2
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-Mail
flask-mongoengine
Flask-RESTful
flask-restplus
flask-swagger-ui
Flask-WTF
google-api-core==1.17.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.18.0
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.8.0
mongoengine==0.19.1
more-itertools==8.5.0
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
protobuf==3.11.3
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.2
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
uWSGI==2.0.19.1
Werkzeug==1.0.1
WTForms==2.2.1
zipp==3.1.0
